requirements.txt
# ...
# Core Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Data Processing
numpy==1.24.3
pandas==2.0.3

# Image Processing
Pillow==10.0.0
opencv-python==4.9.0.80

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning
scikit-learn==1.3.0

# Explainability
tf-keras-vis==0.8.5

# Interactive Demo
gradio==3.41.2

# Utilities
tqdm==4.66.1

# Jupyter
jupyter==1.0.0
notebook==7.2.2
ipykernel==6.25.1
IPython==7.34.0  
typing-extensions==4.5.0 

# Presentation
python-pptx==0.6.21
# ...

